+++ almost_make/utils/shellUtil/runner.py
@@ -41,15 +41,20 @@ def getEndingCount(text, char):
 
     return count
 
-# Remove empty strings from a string array.
-def removeEmpty(array):
+# Return a shallow copy of [array] with 
+# all items equal to [eqlTo] removed.
+def removeEqual(array, eqlTo):
     result = []
 
     for val in array:
-        if val != "":
+        if val != eqlTo:
             result.append(val)
     
     return result
+
+# Remove empty strings from a string array.
+def removeEmpty(array):
+    return removeEqual(array, '')
 
 # Get the number of pairs of parentheses that
 # surround the entirety of [splitText].
@@ -346,7 +351,8 @@ def evalCommand(orderedCommand, customCommands={}, flags=[], stdin=None, stdout=
     else:
         raise SyntaxError("Too many parts to expression, %s" % str(orderedCommand))
 
-# Like shlex.split with punctuation grouping, but groups punctuation intelligently.
+# Like shlex.split but preserves quotation marks. 
+# Send output to filterSplitList for punctuation grouping, but groups punctuation intelligently.
 def shSplit(text, splitChars={ '(', ')', '|', '&', '>', ';', ' ', '\t', '\n' }, quoteChars = { '"', "'" },
         openingParen='(', closingParen=')'):
     escaped = False